Changed the DEFAULT_COMPUTE_API_VERSION value in the constants.py
The default value defined for the variable DEFAULT_COMPUTE_API_VERSION
in the ~/.overcloudrc file generated by overcloudrc.py, which pulls
this variable value from constants.py, is '2.latest', but this
generates a warning message everytime a user uses this standard config
file, like this one below:

/usr/lib/python3.9/site-packages/openstack/config/cloud_region.py:452:
UserWarning: You have a configured API_VERSION with 'latest' in it.
In the context of openstacksdk this doesn't make any sense.
  warnings.warn(

To avoid that, I changed the value '2.latest' to simply '2.1',
which is the current API version in use after a fresh install.

Change-Id: Idbcfbf29626d9a66427b78067813915caf0e9cd5
Signed-off-by: Alexon Oliveira <alolivei@redhat.com>